+++ package.json
@@ -7,7 +7,8 @@
     "immutability-helper": "^3.0.1",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
-    "react-scripts": "3.0.1"
+    "react-scripts": "3.0.1",
+    "react-z-index": "^3.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
